// backend/types.ts
export enum CellType {
    Water = 0,
    FakeWater = 1,
    Tunnel = 2,
    PointerDiamondHorizontal = 3,
    PointerDiamondVertical = 4,
    PointerTurbineHorizontal = 5,
    PointerTurbineVertical = 6,
    PointerEngineHorizontal = 7,
    PointerEngineVertical = 8,
    PointerCompassHorizontal = 9,
    PointerCompassVertical = 10,
    Storm = 11,
    FlightSquare = 12,
    UsualLocation = 13,
    StartPoint = 14
}

export enum Direction {
    Right = 0,
    RightDown = 1,
    Down = 2,
    LeftDown = 3,
    Left = 4,
    LeftUp = 5,
    Up = 6,
    RightUp = 7
}

export enum Character {
    Archeologist = 0,
    WaterKeeper = 1,
    Meteorologist = 2,
    Navigator = 3,
    Explorer = 4,
    Scout = 5
}